# api/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/account": {
            "get": {
                "description": "\u003cb\u003e⚠️ The user must be logged in ⚠️\u003c/b\u003e\u003cbr\u003e\u003cbr\u003eReads the user's account data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Reads the user's account data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/account.UserReadResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "\u003cb\u003e⚠️ The user must be logged in ⚠️\u003c/b\u003e\u003cbr\u003e\u003cbr\u003eUpdates the user's account",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Updates the user's account",
                "parameters": [
                    {
                        "description": "New arguments to be set in the user",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.UpdateArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.PostLoginResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "\u003cb\u003e⚠️ The user must be logged in ⚠️\u003c/b\u003e\u003cbr\u003e\u003cbr\u003eDeletes the user's account from database, the user will also be logged out\u003cbr\u003e \u003cb\u003eCareful, there is no turn back or check, once called, this route will delete the user no matter what\u003c/b\u003e \u003cbr\u003e It also deletes all content related to the user",
                "tags": [
                    "account"
                ],
                "summary": "Deletes the user's account",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "description": "Connection of a user to an existing account\u003cbr\u003eToken will be set in cookies if the arguments combination is valid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Connection of a user to an existing account",
                "parameters": [
                    {
                        "description": "Email + Password combination of the corresponding user",
                        "name": "creds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.Creds"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.PostLoginResponse"
                        }
                    }
                }
            }
        },
...
        "/api/auth/oauth": {
            "post": {
                "description": "Connection of a user using an OAuth \u003cbr\u003eToken will be set in cookies if the arguments combination is valid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Connection of a user",
                "parameters": [
                    {
                        "description": "Code + Service combination of the corresponding OAuth",
                        "name": "creds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.OauthArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structs.PostResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "description": "Connection of a new user to a new account\u003cbr\u003eToken will be set in cookies if the arguments combination is valid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Connection of a new user to a new account",
                "parameters": [
                    {
                        "description": "User related informations that will be later needed for the login process",
                        "name": "creds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.PostLoginResponse"
                        }
                    }
                }
            }
        },
...
        "/api/users": {
            "get": {
                "description": "⚠️ Only accessible to admins ⚠️\u003cbr\u003e\u003cbr\u003eReads all users accounts data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Reads all users data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/users.UserReadResponse"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "⚠️ Only accessible to admins ⚠️\u003cbr\u003e\u003cbr\u003eUpdates the given user's account data",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Updates given ID user with arguments in body",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the user to do the action on",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "New arguments to be set in the user given, along with its id",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.UpdateArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "description": "⚠️ Only accessible to admins ⚠️\u003cbr\u003e\u003cbr\u003eCreates a user account with given informations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Connection of a new user to a new account",
                "parameters": [
                    {
                        "description": "User related informations that will be later needed for the login process",
                        "name": "creds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.CreateUsersArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structs.PostResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "⚠️ Only accessible to admins ⚠️\u003cbr\u003e\u003cbr\u003eDeletes the given user's account\u003cbr\u003e \u003cb\u003eCareful, there is no turn back or check, once called, this route will delete the user no matter what\u003c/b\u003e \u003cbr\u003e It also deletes all content related to the user",
                "tags": [
                    "users"
                ],
                "summary": "Deletes given ID user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the user to do the action on",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/users/id": {
            "get": {
                "description": "⚠️ Only accessible to admins ⚠️\u003cbr\u003e\u003cbr\u003eReads the given user's account data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Reads given ID user data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the user to do the action on",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.UserReadResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "account.UpdateArgs": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "account.UserReadResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "auth.Creds": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.OauthArgs": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                }
            }
        },
        "auth.PostLoginResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterArgs": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "structs.PostResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "users.CreateUsersArgs": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.UpdateArgs": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "users.UserReadResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        }
    }
}